+++ docs/ReleaseNotes.md
@@ -3,6 +3,7 @@
 ## [2.4.0](https://www.nuget.org/packages/WeihanLi.Npoi/2.4.0)
 
 - Fixes <https://github.com/WeihanLi/WeihanLi.Npoi/issues/146>, fix csv encoding handling issue, thanks @yesyeey for spotting the issue
+- `CsvHelper` enhancements
 
 ## [2.3.0](https://www.nuget.org/packages/WeihanLi.Npoi/2.3.0)